+++ src/system/type/String.h
@@ -18,7 +18,7 @@ extern String new_String_sized(uns_int size);
 
 extern String String_concat_(String str1, String str2);
 extern String new_String_from_charp(const char * input);
-extern int Words_compare(Symbol s1, Symbol s2);
+extern long Words_compare(Symbol s1, Symbol s2);
 extern char* unicode_to_ascii(const wchar_t* str);
 extern wchar_t * ascii_to_unicode(const char* str);
